+++ model/Product.cpp
@@ -3,14 +3,14 @@
 #include <limits>
 using namespace std;
 
-Product::Product(string newM,string newNM,double newP):marca(newM),nomeModello(newNM),prezzo(newP){
+Product::Product(string newM,string newNM,float newP):marca(newM),nomeModello(newNM),prezzo(newP){
     if(prezzo<0.0) setPrezzo(0.0);
     else setPrezzo(round(prezzo*100)/100);
 };
 
 //operatori
 bool Product::operator==(const Product& p){
-    return (marca==p.getMarca() && nomeModello==p.getNomeModello() && (fabs(prezzo - p.getPrezzo()) < numeric_limits<double>::epsilon()));
+    return (marca==p.getMarca() && nomeModello==p.getNomeModello() && (fabsf(prezzo - p.getPrezzo()) < numeric_limits<float>::epsilon()));
 }//==
 
 bool Product::operator!=(const Product& p){
@@ -41,7 +41,7 @@ string Product::getNomeModello() const{
     return nomeModello;
 }
 
-double Product::getPrezzo() const{
+float Product::getPrezzo() const{
     return prezzo;
 }//getPrezzo
 
@@ -53,7 +53,7 @@ void Product::setNomeModello(string newNM){
     nomeModello=newNM;
 }//setNomeModello
 
-void Product::setPrezzo(double newP){
+void Product::setPrezzo(float newP){
     prezzo=newP;
 }//setPrezzo
 
@@ -73,6 +73,6 @@ void Product::XML(QXmlStreamWriter &x) const{
     x.writeCharacters(QString::fromStdString(Product::getNomeModello()));
     x.writeEndElement();
     x.writeStartElement("Prezzo");
-    x.writeCharacters(QString::number(Product::getPrezzo()));
+    x.writeCharacters(QString::number((static_cast<double>(Product::getPrezzo()))));
     x.writeEndElement();
 }//xml
